split welcome controller into locale (server) and welcome (client)
