+++ subiquity/client/controllers/__init__.py
@@ -14,7 +14,9 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 from .progress import ProgressController
+from .welcome import WelcomeController
 
 __all__ = [
     'ProgressController',
+    'WelcomeController',
 ]
