+++ subiquity/common/apidef.py
@@ -15,7 +15,7 @@
 
 from typing import List, Optional
 
-from subiquity.common.api.defs import api
+from subiquity.common.api.defs import api, simple_endpoint
 from subiquity.common.types import (
     ApplicationState,
     ApplicationStatus,
@@ -28,6 +28,7 @@ from subiquity.common.types import (
 @api
 class API:
     """The API offered by the subiquity installer process."""
+    locale = simple_endpoint(str)
 
     class meta:
         class status:
