+++ subiquity/controllers/__init__.py
@@ -30,7 +30,6 @@ from .reporting import ReportingController
 from .snaplist import SnapListController
 from .ssh import SSHController
 from .userdata import UserdataController
-from .welcome import WelcomeController
 from .zdev import ZdevController
 
 __all__ = [
@@ -52,6 +51,5 @@ __all__ = [
     'SnapListController',
     'SSHController',
     'UserdataController',
-    'WelcomeController',
     'ZdevController',
 ]
