+++ subiquity/server/controllers/locale.py
@@ -16,15 +16,16 @@
 import logging
 import os
 
-from subiquity.controller import SubiquityTuiController
-from subiquity.ui.views.welcome import get_languages, WelcomeView
-from subiquity.ui.views.help import get_global_addresses
+from subiquity.common.apidef import API
+from subiquity.server.controller import SubiquityController
 
 
-log = logging.getLogger('subiquity.controllers.welcome')
+log = logging.getLogger('subiquity.server.controllers.locale')
 
 
-class WelcomeController(SubiquityTuiController):
+class LocaleController(SubiquityController):
+
+    endpoint = API.locale
 
     autoinstall_key = model_name = "locale"
     autoinstall_schema = {'type': 'string'}
@@ -40,36 +41,7 @@ class WelcomeController(SubiquityTuiController):
         lang = os.environ.get("LANG")
         if lang is not None and lang.endswith(".UTF-8"):
             lang = lang.rsplit('.', 1)[0]
-        for code, name in get_languages():
-            if code == lang:
-                self.model.switch_language(code)
-                break
-        else:
-            self.model.selected_language = lang
-
-    def make_ui(self):
-        language = self.model.selected_language
-        serial = self.app.opts.run_on_serial
-        if serial:
-            ips = get_global_addresses(self.app)
-        else:
-            ips = None
-        return WelcomeView(self, language, serial, ips)
-
-    def run_answers(self):
-        if 'lang' in self.answers:
-            self.done(self.answers['lang'])
-
-    def done(self, code):
-        log.debug("WelcomeController.done %s next_screen", code)
-        self.signal.emit_signal('l10n:language-selected', code)
-        self.model.switch_language(code)
-        self.configured()
-        self.app.next_screen()
-
-    def cancel(self):
-        # Can't go back from here!
-        pass
+        self.model.selected_language = lang
 
     def serialize(self):
         return self.model.selected_language
@@ -79,3 +51,10 @@ class WelcomeController(SubiquityTuiController):
 
     def make_autoinstall(self):
         return self.model.selected_language
+
+    async def GET(self) -> str:
+        return self.model.selected_language
+
+    async def POST(self, data: str):
+        self.model.switch_language(data)
+        self.configured()
